Add Backend Cache
The cache caches not just reads but also writes, for backends that
support it. A backend which returns an object with a `write` method on
an initial write or read will receive subsequent writes to that object
rather than to the model class itself. A model may thus wrap a
translation with something that accepts the `write`, potentially
bypassing costly lookup operations on subsequent writes.