Add Table backend
This is the default Mobility backend. It is based on the Globalize idea
of storing translations in a separate table, but rather than using one
table per model, the backend uses a shared table with key/value columns
corresponding to attribute/value pairs for each locale. The relation
between models and their translations is polymorphic, where a
translation belongs to a "translatable" model.

The major advantage of this approach is that no migrations are required
to add translations to models. Managing translations may also be easier
with all translations stored in the same table.