Add BackendResetter to reset model changes
Both backend cache and backend dirty modules need to be reset on certain
actions, which for now we will limit to `changes_applied` and
`clear_changes_information`, since these are the methods used in
ActiveModel::Dirty. ActiveRecord records mess with the original AM
methods a bit, so we need to monkeypatch them to work in case we have a
non-persisted backend (not currently a tangible thing, but we don't want
to limit use-cases).

Non AR/AM classes will not get reset since we have no way of knowing
where to hook in to reset them.