Only override `where` once, not once per backend include
Currently `where` gets overridden multiple times on the mobility scope
if attributes are split across multiple `translates` calls in the model.
Although this does not cause any problems, we can refactor so we only
override the method once, by checking to see if `MobilityWhereChain` has
already been defined on the model; if it has not been defined, then we
override `where` just once.