Only set alias for mobility_accessor if set in configuration
Many gems use the class method `translates` to define translated
attribute accessors. To avoid possible conflicts, let's define a
configuration option, which defaults to `translates`, to define an alias
for `mobility_accessor`. If an application uses another gem like
Globalize, this option can be set to another name or to a falsy value to
avoid a method name conflict if both gems are enabled for the same
class.