Get sequel serialized backend to work with update
There seems to be an issue with the sequel serialization modification
detection plugin, whereby the hash tracking the original deserialized
values points to the same object as the deserialized values hash, so
that changes to the latter are the same as the former. In our case this
prevents the plugin from really working.

I changed the code to do a depth-2 dup so that values are destinct,
which fixes the failing specs.