Ensure model updates when Sequel table backend translations updated
In order to ensure that changes are tracked on the model even for
"virtual" translated columns, we override mobility_set to add a column
to changed_columns for the translated attribute. This is an easy and
quick way to ensure that Sequel knows the model has been modified, and
thus should save it.