Mark both attr and attr_<locale> in column changes module
If we do not add the attribute name itself to column_changes, then
backends that actually change the a column with the name of the
attribute will not get the column updated, since Sequel checks to see if
the changed column is actually a real column.

So instead, we should mark *both* e.g. `title` *and* `title_en`.