Add AR table backend
The table backend uses a translation table for each model to store
translated attributes, in essentially the same way that Globalize does.
The implementations are not identical however, and in particular
Mobility does not add many of the methods that Globalize adds to model
classes to access model translations. The basic accessor and querying
interfaces, however, are mostly the same.