Commpletely isolate attributes in table backend
In line with the general philosophy of Mobility, this change isolates
attributes defined using the table backend so they can each potentially
use their own table. The association_name must be different for each in
order for this to work.

This then makes it possible to have a model, Post, with a table `posts`,
which has translations for attributues spread across multiple tables.
If title is on `post_title_translations` and content is on
`post_content_translations`, you would use:

class Post < ActiveRecord::Base
  include Mobility
  translates :title, backend: :table, association_name: :title_translations, table_name: :post_title_translations
  translates :content, backend: :table, association_name: :content_translations, table_name: :post_content_translations
end

Note that the `association_name` here is arbitrary; as long as each
backend has *different* names for the association, this will work.

This commit also removes methods like `mobility_translation_for` from
the model itself, and moves them to the backend, limiting sources of
conflict in the case of multiple backends with different tables (above).