Include fallbacks module by default but do not actually use it
It is now possible to set fallbacks through options on the translated
attribute getter method (`post.title(fallback: :de)`), but if the
fallbacks module is not included then this will not work. Currently, by
default `Mobility::Backend::Fallbacks` is not included by default.

This changes the setup so that we include the module by default, but
if the value in the backend options hash is `nil`, then we do not use it
except in the case above. This should be more intuitive.