Inject double into backend class for testing
We want to be able to stub methods on the backend instance so we can see
what is actually getting passed through to it, but the class itself will
get subclassed and include modules (cache, fallbacks, etc) which will
make it hard to get at the original read/write methods. We get around
this by injecting a double into the backend so we can stub methods on
the double in specs.