Destroy all translations associated with model when it is destroyed
The KeyValue backend uses a polymorphic association where key columns
store attributes, and we limit the association to only the keys defined
as translated attributes for the model (for performance, etc.) This has
the consequence that although we may have a dependent: destroy on the
model, translations which were previously stored with different keys
(for example, if we changed attribute names in the model) will be left
behind, because they are not seen by the association (anymore).

Having them left behind is unavoidable, but at minimum we should destroy
them all when the model is destroyed to avoid them begin associated with
other, new models. To do this we add an after destroy callback which
destroys *all* string and text translations associated with the model,
regardless of their attribute name (key). This ensures that when we
destroy a translated record, we are guaranteed that the entire record of
the record is wiped from the database.