Make model_class not required
Previously model_class was set from the Translates module when the class
method `translates` was called from a model, but we should not require
this. Instead, we can do most of the model setup in the `includes`
method where we already know the model class, so we don't need it when
creating the `Attributes` module instance anymore.