Only include plugins in default_options if their default is truthy
It's more confusing than anything to have a bunch of stuff in default
options that actually has no impact on anything. The only defaults in
default_options that matter are cache and presence, which are turned on
by default; the other ones can be nil and nothing changes, so we can
just leave them out.

If someone genuinely wants to entirely disable a plugin globally, the
way to do it is by removing it from the +plugins+ array.