Replace QueryMethods with nodes and Query plugin
I've left this as one giant commit, because to split it is more work
that it is worth. This change replaces module builders for each backend
with a single plugin which delegates to the methods build_node (which
returns an Arel node) and (optionally) add_translations (which modifies
an AR::Relation) on the backend class.

Backend classes now need only to implement these one or two methods and
querying is supported without any extra work.