Apply scope to relation + predicate in backend class
Previously, a backend class defined a method +add_translations+ which
took a relation, an options hash, a locale and an optional invert
boolean and applied a scope to the relation (used by Table and KeyValue
backends to join their translations).

This commit changes the method name to +add_scope+ and instead of taking
an options hash, takes an Arel node (a predicate) which it then
(optionally) visits to apply any extra changes to the relation. This
allows backends to apply much more fine-grained changes in a more
consistent way, since the predicate can be any query whereas an options
hash is limited to equality/inequality predicates.

The commit also adds visitor classes for Table and KeyValue backends
which are used in +apply_scope+ to add INNER and/or OUTER joins to a
relation depending on the predicate passed as an argument.

To handle this new format more effectively, the query method +i18n+ now
optionally takes a block which it evaluates in the context of a
VirtualRow class. This makes it easy to execute complex queries, e.g.

  Post.i18n { title.eq("foo").or(content.eq("bar")) }

Any translated attribute can be used as a bare word in this block. In
addition, normal columns (untranslated attributes) can also be used.