Update readme with info on querying
Adding some info on the new block-based interface for querying. If
anyone is interested, please checkout the master branch and give this a
go. I will release it in the next week or two, probably ActiveRecord
alone first (not yet sure how Sequel will come out).

This will be very useful for gem integrations in particular, since it
means that a gem that involves querying can be integrated once for all
backends. This has been done already for FriendlyId without this level
of granularity because querying in FriendlyId is quite simple, but most
other querying gems require a more fine-grained interface. Ransack is
the one I'm thinking of integrating first.

Within Mobility, the first concrete benefit of this new querying
interface is that it has enabled case insensitive uniqueness validation
(which requires `LOWER` on an Arel predicate, now possible thanks to #235.

[ci skip]