Check rails i18n locales when enforcing available locales
This is a subtle edge-case, but I18n.locale_available? will not actually
check that the current locale is included in *rails-defined* locales.
This may cause problems, for example, if you set available locales in
Rails and then try to use Mobility in an initializer, since Rails may
not yet have set I18n.available_locales.