Define custom mutation tracker for AR/AM dirty plugins
In the past, Mobility has attempted to leverage ivar-backed
dirty-tracking in ActiveModel and ActiveRecord, but this approach has
become unsustainable since Rails has never openly supported this in
ActiveRecord and now basically does not support it at all.

Instead, we define a custom mutation tracker and override/define all
necessary dirty methods explicitly. This requires more lines of code,
but in the end it is easier to understand and reason about, and should
be more robust to changes in Rails.

By inlining the mutation tracker instead of trying to inherit from
mutation tracker classes in ActiveModel, we are able to support all
versions of Rails going as far as 4.2 with no conditionals.