Ensure that public/private scope of changes_applied, etc do not change
Previously, we were looking at the Active*Model* methods to se if they
were private or public, and defining overrides accordingly. However, in
some versions of Rails, the methods can be private in one module, but
public in the other... so we need to ensure that the code actually looks
at the correct +dirty_class+ to determine if the methods should be
public or private.