Avoid making private methods accidentally public
`changes_applied` is private in some versions of Rails, and we should
not be changing that. To avoid this from happening, we can create a
dummy class including the AR dirty module, check its methods and see if
either `changes_applied` or `clear_changes_information` are private, and
if so change the method visibility accordingly.