Specify plugins in configure block with before/after dependencies
Currently, plugins are declared inside the body of a
Mobility::Attributes instance. Dependencies declared with depends_on in
the plugin are then included after the plugin itself.

This may not always be what we want, since multiple plugins may depend
on each other. To correctly resolve dependencies, this requires wrapping
plugin declarations in a block:

  class TranslatedAttributes < Mobility::Attributes
    plugins do
      foo
      bar
    end
  end

In addition, plugins now declare dependencies with a required keyword
argument, include: :before or include: :after, which specifies whether
to include the dependency before or after the plugin itself. We then use
TSort to sort dependencies such that the requirements on include order
are met.