Register plugins explicitly
This is an idea borrowed from Shrine, Roda and other gems. Basically, we
shouldn't be just assuming every plugin is in the Mobility::Plugins
namespace, and that everything in this namespace is a plugin (some
classes like Mobility::Plugins::ActiveModel, etc are actually not).

By explicitly declaring the mapping between keys and plugin modules,
things are clearer, easier to test, and easier to extend (e.g. if you
don't require a plugin from Mobility, you can use the same key for
your own plugin module).