Improve dependency conflict error messages and simplify create_tree
The most natural way to define plugins is:

class TranslatedAttributes < Mobility::Attributes
  plugin :foo
  plugin :bar
end

However, there is an issue here if bar needs foo to be included after
itself. In this case, Mobility will raise a CyclicDependency error,
which is not semantically the correct error to raise since there isn't
really a cyclic dependency; the fix is just to declare 'foo' after
'bar'.

With this change, we simply detect such issues when they occur and raise
the appropriate error. This also simplifies the dependency resolution
logic because we don't need to add all previously included plugin
modules when building the dependency tree in create_tree.