Introduce Mobility::Pluggable as abstract parent for Mobility::Attributes
Plugin inclusion and dependency resolving logic is pretty much entirely
divorced now from any real concept of "attributes", so we can split the
two cleanly and just have Mobility::Attributes subclass a minimal Module
subclass, which I'll call Mobility::Pluggable.

With this separation, the remaining methods in Mobility::Attributes are
non-essential and I think we can simply remove most of them since they
are not widely used. The goal here is to have as close to zero
assumptions baked into Attributes as possible.