Add BackendReader
The Backend plugin does three basic things on the model class:
- defines getter method for each attribute (`title_backend` etc)
- defines `mobility_backends` (instance method) and
  `mobility_backend_class` (class method) on model class
- sets up model with backend (calls backend.setup_model on model)

The defining of getter methods is not really required for any other
plugins, since internally everything goes to `mobility_backends[:title]`
instead of `title_backend` (former is faster, and for accessor methods
in particular speed is important).

Also, the "X_backend" format being hard-coded has always kind of
bothered me. So I thought, now that we have plugins, why not just split
off that part into its own plugin? Call it BackendReader. And the format
string can be the argument to that plugin, since it doesn't actually
need to know the backend itself to define the reader method.