Set plugin default via DSL instead of keyword argument
Setting the default for a plugin via a keyword argument to
initialize_hook and/or included_hook is problematic because then other
plugins cannot see this default. It's better just to be explicit about
it and declare the default, then we can set it on the defaults hash as
soon as the plugin is included into a pluggable (module), so all plugins
can see it from both hooks.