Remove unnecessary const_set for uniqueness validator
I noticed that this const_set here is not actually doing anything, and
in fact uniquenes validation works simply because namespacing leads
`conts_get` to `Mobility::ActiveRecord::UniquenessValidator` before
`ActiveRecord::Validations::UniquenessValidator`.

This is probably not good... we'd prefer to have a new subclass then to
magically resolve to the Mobility validator class. Need to figure out a
better way.