Do not register backend classes that are not actually backends
This is the case where e.g. you set a default backend, but do not
include either active_record or sequel plugins, so we fall through to
the base +Mobility::Backends::KeyValue+ class, which is not actually a
backend class.

This leads to weird results. Better to raise if the ORM plugin is
missing.