Call configure on subclass with model_class in context
Merging the model class into the options isn't really a very good idea,
since it clutters the hash and makes it less clear what comes from
where. Instead, we can create the subclass first, define the model_class
method so that when `configure` is called, it has the model class in its
context and doesn't need it to be added to the options hash.