Load class before setting up model
The table backend will create a class of the form Article::Translation,
but if the constant Article has not yet been loaded, strange things
happen.

As it turns out, this `let` block does not necessarily load the constant
since it is sometimes passed in as a class and assigned with stub_const,
without the two being yet connected together.

To get around this, we constantize the `model_class.name` here so the
two match.