Replace apply_plugin with include_cache
The name +apply_plugin+ falsely indicates that this is a generic method,
when in fact it has a very specific use case in the cache plugin. Let's
remove it and just make this explicit, so now the cache plugin looks for
a method on the backend, which if present will be called, otherwise
defaults to simple cache implementation.

This is not beautiful but I think better since it reduces clutter and
complexity in the base Mobility::Backend module.