Clearly distinguish backend classes from their configured subclasses
In practice, every backend class that is used in a model is a subclass
of one of the base classes. Configured backend classes have a model
class and set of options, whereas the base backend classes do not.

However, previously you could still call `options` or `model_class` on
the base classes and you would get back a `nil`. This is confusing since
those getter methods should never be called on the base backend classes.

In this commit I'm restructuring the code such that:
- getter methods raise on the base backend class
- methods which do not make sense on the base backend class are moved
  to a module which is included into backend subclasses.

I think this should avoid weird bugs and misunderstandings.